requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements.in
#
click==8.0.1
    # via w1thermsensor
importlib-metadata==4.8.1
    # via click
prometheus-client==0.11.0
    # via -r requirements.in
python-dotenv==0.19.0
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via importlib-metadata
w1thermsensor==2.0.0
    # via -r requirements.in
zipp==3.5.0
    # via importlib-metadata
